Add feature flag for desktop solution
